transport: drop locator/tablets.hh and replica/tablets.hh from result_message_base.hh
result_message_base.hh is included by ~213 TUs. It previously included
locator/tablets.hh (which pulls in seastar/reactor.hh, raft/raft.hh, etc.),
replica/tablets.hh, and types/tuple.hh solely to support the inline body
of add_tablet_info().

Move add_tablet_info() out of line into result_message.cc. Introduce
locator/tablets_fwd.hh with the lightweight types tablet_replica,
tablet_replica_set, and tablet_routing_info (the last one was added to
master after the original patch). The header now only needs tablets_fwd.hh
and bytes_fwd.hh.

(cherry picked from commit e642651899e35db8d7af5b88eacbd3e84223912a,
originally part of scylladb/scylladb#29009)
