+++ locator/tablets.hh
@@ -8,10 +8,9 @@
 
 #pragma once
 
+#include "locator/tablets_fwd.hh"
 #include "dht/token.hh"
 #include "locator/token_metadata_fwd.hh"
-#include "utils/small_vector.hh"
-#include "locator/host_id.hh"
 #include "service/session.hh"
 #include "dht/i_partitioner_fwd.hh"
 #include "dht/token-sharding.hh"
@@ -105,15 +104,6 @@ struct range_based_tablet_id {
 
     bool operator==(const range_based_tablet_id&) const = default;
 };
-
-struct tablet_replica {
-    host_id host;
-    shard_id shard;
-
-    auto operator<=>(const tablet_replica&) const = default;
-};
-
-using tablet_replica_set = utils::small_vector<tablet_replica, 3>;
 
 /// A 64-bit hash computed from a tablet's replica list after ordering it.
 /// Used by TABLETS_ROUTING_V2 to detect when a driver's cached routing is stale.
@@ -1076,11 +1066,6 @@ public:
 future<bool> check_tablet_replica_shards(const tablet_metadata& tm, host_id this_host);
 
 std::optional<tablet_replica> maybe_get_primary_replica(tablet_id id, const tablet_replica_set& replica_set, const locator::topology& topo, std::function<bool(const tablet_replica&)> filter);
-
-struct tablet_routing_info {
-    tablet_replica_set tablet_replicas;
-    std::pair<dht::token, dht::token> token_range;
-};
 
 struct tablet_routing_info_v2 {
     tablet_replica_set tablet_replicas;
