+++ transport/messages/result_message.cc
@@ -8,10 +8,11 @@
  */
 
 #include "result_message.hh"
-#include "cql3/cql_statement.hh"
+#include "locator/tablets.hh"
 #include "replica/tablets.hh"
 #include "types/tuple.hh"
 #include <bit>
+#include "cql3/cql_statement.hh"
 #include <seastar/core/format.hh>
 #include <fmt/std.h>
